+++ bundles/org.palladiosimulator.dataflow.confidentiality.analysis/META-INF/MANIFEST.MF
@@ -20,7 +20,6 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="3.24.0",
  org.palladiosimulator.mdsdprofiles.api;bundle-version="4.3.0";visibility:=reexport,
  org.palladiosimulator.dataflow.nodecharacteristics;bundle-version="0.1.0"
 Export-Package: org.palladiosimulator.dataflow.confidentiality.analysis,
- org.palladiosimulator.dataflow.confidentiality.analysis.dsl,
- org.palladiosimulator.dataflow.confidentiality.analysis.sequence.entity,
- org.palladiosimulator.dataflow.confidentiality.analysis.sequence.entity.pcm,
+ org.palladiosimulator.dataflow.confidentiality.analysis.entity,
+ org.palladiosimulator.dataflow.confidentiality.analysis.entity.pcm,
  org.palladiosimulator.dataflow.confidentiality.analysis.sequence.pcm
