+++ bundles/org.palladiosimulator.dataflow.confidentiality.analysis/src/org/palladiosimulator/dataflow/confidentiality/analysis/characteristics/NodeCharacteristicsCalculator.java
@@ -1,10 +1,10 @@
-package org.palladiosimulator.dataflow.confidentiality.analysis.sequence.pcm.node;
+package org.palladiosimulator.dataflow.confidentiality.analysis.characteristics;
 
 import java.util.Deque;
 import java.util.List;
 import java.util.Optional;
 
-import org.palladiosimulator.dataflow.confidentiality.analysis.sequence.entity.CharacteristicValue;
+import org.palladiosimulator.dataflow.confidentiality.analysis.entity.CharacteristicValue;
 import org.palladiosimulator.pcm.core.composition.AssemblyContext;
 
 public interface NodeCharacteristicsCalculator {
