--- bundles/org.palladiosimulator.dataflow.confidentiality.analysis/src/org/palladiosimulator/dataflow/confidentiality/analysis/dsl/DataFlowAnalysis.java
@@ -1,5 +1,0 @@
-package org.palladiosimulator.dataflow.confidentiality.analysis.dsl;
-
-public final class DataFlowAnalysis {
-    // TODO: Implement DSL
-}
