+++ bundles/org.palladiosimulator.dataflow.confidentiality.analysis/src/org/palladiosimulator/dataflow/confidentiality/analysis/entity/ActionSequence.java
@@ -1,4 +1,4 @@
-package org.palladiosimulator.dataflow.confidentiality.analysis.sequence.entity;
+package org.palladiosimulator.dataflow.confidentiality.analysis.entity;
 
 import java.util.List;
