+++ bundles/org.palladiosimulator.dataflow.confidentiality.analysis/src/org/palladiosimulator/dataflow/confidentiality/analysis/entity/CallReturnBehavior.java
@@ -1,4 +1,4 @@
-package org.palladiosimulator.dataflow.confidentiality.analysis.sequence.entity;
+package org.palladiosimulator.dataflow.confidentiality.analysis.entity;
 
 public interface CallReturnBehavior {
